refactor: Application 클래스에 몰려있던 게임 진행 로직을 PlayGroundManager 클래스로 분리
기존에는 Application 클래스에서 사용자 입력에 대한 처리와 게임 진행을
같이 했기 때문에 소스 코드가 복잡한 느낌이 있었다. 이를 해결하기 위해
PlayGround의 동작을 제어하는 PlayGroundManager 클래스를 새로 생성하고
Application 클래스에 있던 게임 진행 로직은 모두 PlayGroundManager
클래스로 이관하였다.
