+++ src/main/java/baseball/InputParser.java
@@ -3,9 +3,20 @@ package baseball;
 import java.util.ArrayList;
 
 public class InputParser {
-    public static final String INPUT_SPLIT_DELIMITER = "";
+    private static final String INPUT_BALL_NUMBER_ERROR_MSG = "[ERROR] 입력한 값이 유효하지 않습니다. 1~9 사이의 정수값 3개를 입력해주세요...";
+    private static final String INPUT_SPLIT_DELIMITER = "";
 
     public ArrayList<Integer> parse(String input) {
+        try {
+            return this.parseInternal(input);
+        } catch (IllegalArgumentException e) {
+            System.out.println(INPUT_BALL_NUMBER_ERROR_MSG);
+
+            return new ArrayList<>();
+        }
+    }
+
+    private ArrayList<Integer> parseInternal(String input) {
         String[] splits = input.split(INPUT_SPLIT_DELIMITER);
 
         verifyInputLength(splits);
